change of development environment

+++ include/SegTable.h
@@ -6,9 +6,9 @@
 typedef struct SegTable {
   char * ptr;
   Dimensions dim;
-  int length;
+  short length;
 } SegTable;
-SegTable SegTable_init(char * fname, int size, int elementSize){
+SegTable SegTable_init(char * fname, unsigned short size, unsigned short elementSize){
   SegTable s;
   if(size%elementSize==0){
     if(fname!=0){
@@ -26,8 +26,8 @@ SegTable SegTable_init(char * fname, int size, int elementSize){
 void SegTable_free(char * ptr){
   free(ptr);
 }
-int SegTable_add(char * ptr, SegTable table){
-  int i=0;
+short SegTable_add(char * ptr, SegTable table){
+  short i=0;
   if(table.length >= table.dim.hight) return 0;
   while(ptr[i]!=0 && i<table.dim.width){
     table.ptr[(table.length*table.dim.width)+i]=ptr[i];
@@ -36,8 +36,8 @@ int SegTable_add(char * ptr, SegTable table){
   table.length++;
   return i;
 }
-int SegTable_put(char * ptr, SegTable table, int index){
-  int i=0;
+short SegTable_put(char * ptr, SegTable table, short index){
+  short i=0;
   if(index >= table.length) return 0;
   while(ptr[i]!=0 && i<table.dim.width){
     table.ptr[(index*table.dim.width)+i]=ptr[i];
@@ -45,9 +45,9 @@ int SegTable_put(char * ptr, SegTable table, int index){
   }
   return i;
 }
-char * SegTable_get(int index, SegTable table){
+char * SegTable_get(short index, SegTable table){
   char * ptr=calloc(table.dim.width, sizeof(char));
-  for(int i=0; i<table.dim.width; i++){
+  for(short i=0; i<table.dim.width; i++){
     ptr[i]=table.ptr[(index*table.dim.width)+i];
   }
   return ptr;
